% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_time.R
\name{degree_time}
\alias{degree_time}
\title{Calculate time in each degree}
\usage{
degree_time(data, thresholds)
}
\arguments{
\item{data}{data in wide format with minimum temperature labeled as tmin and maximum temperature labeled as tmax}

\item{thresholds}{threshold of temperature intervals to calculate time in each degree}
}
\description{
\code{degree_time} returns a data frame with calculated time in each degree at one degree intervals
within a specified thresholds.
}
\details{
To generate time in each degree the data passed must be in wide format with minimum temperature 
column labeled as tmin and maximum temperature labeled as tmax.
}
\examples{
data(napa)
degree_time(napa, thresholds = c(0:35))

}
