\name{NonlinearTemp-package}
\alias{NonlinearTemp-package}
\alias{NonlinearTemp}
\docType{package}
\title{
\packageTitle{NonlinearTemp}
}
\description{
\packageDescription{NonlinearTemp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NonlinearTemp}
\packageIndices{NonlinearTemp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NonlinearTemp}

Maintainer: \packageMaintainer{NonlinearTemp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
